backend/requirements.txt
# FastAPI 프레임워크
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# 데이터 검증
pydantic>=2.10.0
pydantic-settings>=2.7.0

# 데이터베이스
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9

# 인증/보안
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=44.0.0

# HTTP 클라이언트
httpx>=0.28.0

# 환경 변수
python-dotenv>=1.0.1
python-multipart>=0.0.20

# AI 지원 (선택)
openai>=1.59.0
anthropic>=0.42.0

# 개발 도구
pytest>=8.0.0
pytest-asyncio>=0.24.0
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
